+++ config/initializers/authy.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+Authy.api_key = Tenant.credentials.authy[:api_secret]
+Authy.api_uri = 'https://api.authy.com'
